+++ Sonic12Decomp/Text.cpp
@@ -125,9 +125,9 @@ void LoadConfigListText(TextMenu *menu, int listNo)
 {
     FileInfo info;
     char strBuf[0x100];
-    int fileBuffer = 0;
-    int count      = 0;
-    int strLen     = 0;
+    byte fileBuffer = 0;
+    byte count      = 0;
+    byte strLen     = 0;
     if (LoadFile("Data/Game/GameConfig.bin", &info)) {
         // Name
         FileRead(&strLen, 1);
@@ -144,14 +144,14 @@ void LoadConfigListText(TextMenu *menu, int listNo)
 
         // Object Names
         FileRead(&count, 1);
-        for (int o = 0; o < count; ++o) {
+        for (byte o = 0; o < count; ++o) {
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
             strBuf[strLen] = 0;
         }
 
         // Script Paths
-        for (int s = 0; s < count; ++s) {
+        for (byte s = 0; s < count; ++s) {
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
             strBuf[strLen] = 0;
@@ -159,7 +159,7 @@ void LoadConfigListText(TextMenu *menu, int listNo)
 
         // Variables
         FileRead(&count, 1);
-        for (int v = 0; v < count; ++v) {
+        for (byte v = 0; v < count; ++v) {
             //Var Name
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
@@ -174,13 +174,13 @@ void LoadConfigListText(TextMenu *menu, int listNo)
 
         // SFX Names
         FileRead(&count, 1);
-        for (int s = 0; s < count; ++s) {
+        for (byte s = 0; s < count; ++s) {
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
             strBuf[strLen] = 0;
         }
         // SFX Paths
-        for (int s = 0; s < count; ++s) {
+        for (byte s = 0; s < count; ++s) {
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
             strBuf[strLen] = 0;
@@ -188,7 +188,7 @@ void LoadConfigListText(TextMenu *menu, int listNo)
 
         // Players
         FileRead(&count, 1);
-        for (int p = 0; p < count; ++p) {
+        for (byte p = 0; p < count; ++p) {
             FileRead(&strLen, 1);
             FileRead(&strBuf, strLen);
             strBuf[strLen] = '\0';
@@ -200,10 +200,10 @@ void LoadConfigListText(TextMenu *menu, int listNo)
         }
 
         // Categories
-        for (int c = 1; c <= 4; ++c) {
+        for (byte c = 1; c <= 4; ++c) {
             int stageCnt = 0;
             FileRead(&stageCnt, 1);
-            for (int s = 0; s < stageCnt; ++s) {
+            for (byte s = 0; s < stageCnt; ++s) {
                 //Stage Folder
                 FileRead(&strLen, 1);
                 FileRead(&strBuf, strLen);
